+++ CollisionCommonModule/src/main/java/io/asteroidsjaylib/collisioncommon/CircleColliderComponent.java
@@ -3,5 +3,5 @@ package io.asteroidsjaylib.collisioncommon;
 import io.asteroidsjaylib.common.ecs.BaseComponent;
 
 public class CircleColliderComponent extends BaseComponent {
-    public double radius;
+    public float radius;
 }
