+++ CommonModule/src/main/java/io/asteroidsjaylib/common/ecs/BaseSystem.java
@@ -15,31 +15,25 @@ public abstract class BaseSystem {
         this.setPriority(priority);
     }
 
-    /**
-     * Called once at system startup.
-     * @param world the world the system operates in.
-     */
+    /// Called once at system startup.
+    /// @param world the world the system operates in.
     public abstract void start(World world);
 
     public abstract List<Class<? extends BaseComponent>> getSignature();
-    public abstract void update(World world, List<BaseEntity> entities, double deltaTime);
+    public abstract void update(World world, List<BaseEntity> entities, float deltaTime);
 
-    /**
-     * Gets the priority of this system.
-     * Systems with lower values are processed first.
-     *
-     * @return the priority value
-     */
+    /// Gets the priority of this system.
+    /// Systems with lower values are processed first.
+    ///
+    /// @return the priority value
     public int getPriority() {
         return priority;
     }
 
-    /**
-     * Sets the priority of this system.
-     * Systems with lower values are processed first.
-     *
-     * @param priority the priority value to set
-     */
+    /// Sets the priority of this system.
+    /// Systems with lower values are processed first.
+    ///
+    /// @param priority the priority value to set
     public void setPriority(int priority) {
         this.priority = priority;
     }
