+++ PhysicsCommonModule/src/main/java/io/asteroidsjaylib/physicscommon/AngleComponent.java
@@ -3,6 +3,6 @@ package io.asteroidsjaylib.physicscommon;
 import io.asteroidsjaylib.common.ecs.BaseComponent;
 
 public class AngleComponent extends BaseComponent {
-    /// Angle in radians
-    public double angle;
+    /// Angle in degrees
+    public float angle;
 }
