+++ PhysicsCommonModule/src/main/java/io/asteroidsjaylib/physicscommon/DragComponent.java
@@ -4,5 +4,5 @@ import io.asteroidsjaylib.common.ecs.BaseComponent;
 
 public class DragComponent extends BaseComponent {
     /// The percentage of velocity that remains after 1 second (0.0 = instant stop, 1.0 = no friction).
-    public double drag;
+    public float drag;
 }
