+++ PhysicsCommonModule/src/main/java/io/asteroidsjaylib/physicscommon/RotationComponent.java
@@ -3,6 +3,6 @@ package io.asteroidsjaylib.physicscommon;
 import io.asteroidsjaylib.common.ecs.BaseComponent;
 
 public class RotationComponent extends BaseComponent {
-    /// In radians/sec
-    public double dAngle;
+    /// In degrees/sec
+    public float dAngle;
 }
